+++ README.md
@@ -2,10 +2,7 @@
 
 Simple tool to scan a PHP code to look for potential injection vulnerabilities. Made using nikic/PHP-Parser, written in PHP.
 
-to run: `php fscream.php PATH/TO/YOUR_FILE`
+to run: `php merged.php PATH/TO/YOUR_PROJECT/`
 
+notice the trailing "/". target has to be a directory.
 
-currently,  only check variable nodes for sources and 
-            only checks function nodes for sanitizers & sinks.
-
-also, does not check globally for tainted-clear inputs (between blocks)
